renamed .h files to hpp and cmake stuff
