+++ firmware/.gitignore
@@ -1,4 +1,4 @@
 bin
-build
+build*
 .vscode/*
 !.vscode/tasks.json
