+++ firmware/CMakeLists.txt
@@ -1,10 +1,13 @@
 cmake_minimum_required(VERSION 3.16)
 
-include(cmake/arm.cmake)
+project(zol ASM CXX)
 
-project(zol ASM C CXX)
+set(CMAKE_CXX_STANDARD 20)
+set(CMAKE_CXX_STANDARD_REQUIRED True)
 
-include(cmake/zol.cmake)
+include(arm-cmake-toolchains/utils.cmake)
+include(arm-cmake-toolchains/clang_utils.cmake)
+
 include(chip/stm32l011.cmake)
 
 add_subdirectory(lib)
