+++ firmware/api/chip/lptim.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "register.h"
+#include "register.hpp"
 
 #include <cstdint>
