+++ firmware/api/fields.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "register.h"
+#include "register.hpp"
 
 #include <cinttypes>
 
@@ -13,7 +13,7 @@ namespace zol {
 	struct field_rw {
 		/// @brief Sets the field to value.
 		/// @throws Compile time error if value does not fit into the field.
-		static void set(reg::type_t value) {
+		static void set(typename reg::type_t value) {
 			// static_assert((1 << width) < value,
 			// 			  "value is greater than allowed bit width");
 			reg::and_reg(~(((1 << width) - 1) << start));
@@ -22,14 +22,14 @@ namespace zol {
 
 		/// @brief Get the value of the field
 		/// @returns the value of the field, duh
-		static reg::type_t get() {
+		static typename reg::type_t get() {
 			return reg::get_reg() & (((1 << width) - 1) << start);
 		}
 	};
 
 	template<zol::Register reg, uint8_t start, uint8_t width>
 	struct field_r {
-		static reg::type_t get() {
+		static typename reg::type_t get() {
 			return reg::get_reg() & (((1 << width) - 1) << start);
 		}
 	};
