+++ firmware/app/bat.cpp
@@ -1,3 +1,3 @@
-#include "bat.h"
+#include "bat.hpp"
 
 int16_t battery::adc_buffer[1];
