--- firmware/cmake/zol.cmake
@@ -1,4 +1,0 @@
-set(CMAKE_CXX_STANDARD 20)
-set(CMAKE_CXX_STANDARD_REQUIRED True)
-
-# set(CMAKE_CXX_CLANG_TIDY "clang-tidy;-checks=*")
